2020.2.16 commits

+++ Aliyun_Python/Practice/6_Dog_Life.py
@@ -2,7 +2,7 @@
 # Formula: dog's life = 10.5 * the first two years + 4 * the following years
 # Question: Convert dog's life when user enters the number
 
-life = int(input('Please enter your dog\'s life: '))
+life = float(input('Please enter your dog\'s life: '))
 
 if life < 0:
     print('Please enter a valid number.')
